Create OpenGraphInspector LiveView & OpenGraphComponent
The OpenGraphInspector displays a form with a single url input and a button to submit,
as well as a list of OpenGraphs fetched. When submitting a new url, the application
kicks off a process to fetch the url and gather it's OpenGraph data. While processing,
a loading indicator is displayed. When processing completes, the loading indicator is
removed and a OpenGraphComponent with the newly collected data is loaded in its place.